pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "openai>=1.93.3",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pyaudio>=0.2.14",
    "requests>=2.32.4",
    "scipy>=1.16.0",
    "soundfile>=0.13.1",
    "speechrecognition>=3.14.3",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
]
